# Don't send documentation in completion list responses.