--- a/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/CompletionItemExtensions.cs
+++ b/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/CompletionItemExtensions.cs
@@ -2,8 +2,6 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using Microsoft.CodeAnalysis.Razor.Completion;
-using Microsoft.CodeAnalysis.Razor.Tooltip;
 using Newtonsoft.Json.Linq;
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 
@@ -11,11 +9,9 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
 {
     internal static class CompletionItemExtensions
     {
-        private const string TagHelperElementDataKey = "_TagHelperElementData_";
-        private const string AttributeCompletionDataKey = "_AttributeCompletion_";
-        private const string RazorCompletionItemKind = "_CompletionItemKind_";
+        private const string ResultIdKey = "_resultId";
 
-        public static void SetRazorCompletionKind(this CompletionItem completion, RazorCompletionItemKind completionItemKind)
+        public static void SetCompletionListResultId(this CompletionItem completion, long resultId)
         {
             if (completion is null)
             {
@@ -23,86 +19,25 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
             }
 
             var data = completion.Data ?? new JObject();
-            data[RazorCompletionItemKind] = JToken.FromObject(completionItemKind);
+            data[ResultIdKey] = resultId;
             completion.Data = data;
         }
 
-        public static bool TryGetRazorCompletionKind(this CompletionItem completion, out RazorCompletionItemKind completionItemKind)
+        public static bool TryGetCompletionListResultId(this CompletionItem completion, out int resultId)
         {
             if (completion is null)
             {
                 throw new ArgumentNullException(nameof(completion));
             }
 
-            if (completion.Data is JObject data && data.ContainsKey(RazorCompletionItemKind))
+            if (completion.Data is JObject data && data.ContainsKey(ResultIdKey))
             {
-                completionItemKind = data[RazorCompletionItemKind].ToObject<RazorCompletionItemKind>();
+                resultId = data[ResultIdKey].ToObject<int>();
                 return true;
             }
 
-            completionItemKind = default;
+            resultId = default;
             return false;
-        }
-
-        public static bool IsTagHelperElementCompletion(this CompletionItem completion)
-        {
-            if (completion.Data is JObject data && data.ContainsKey(TagHelperElementDataKey))
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-        public static void SetDescriptionInfo(this CompletionItem completion, AggregateBoundElementDescription elementDescriptionInfo)
-        {
-            var data = completion.Data ?? new JObject();
-            data[TagHelperElementDataKey] = JObject.FromObject(elementDescriptionInfo);
-            completion.Data = data;
-        }
-
-        public static void SetDescriptionInfo(this CompletionItem completion, AggregateBoundAttributeDescription attributeDescriptionInfo)
-        {
-            if (completion is null)
-            {
-                throw new ArgumentNullException(nameof(completion));
-            }
-
-            if (attributeDescriptionInfo is null)
-            {
-                throw new ArgumentNullException(nameof(attributeDescriptionInfo));
-            }
-
-            var data = completion.Data ?? new JObject();
-            data[AttributeCompletionDataKey] = JObject.FromObject(attributeDescriptionInfo);
-            completion.Data = data;
-        }
-
-        public static AggregateBoundElementDescription GetElementDescriptionInfo(this CompletionItem completion)
-        {
-            if (completion.Data is JObject data && data.ContainsKey(TagHelperElementDataKey))
-            {
-                var descriptionInfo = data[TagHelperElementDataKey].ToObject<AggregateBoundElementDescription>();
-                return descriptionInfo;
-            }
-
-            return AggregateBoundElementDescription.Default;
-        }
-
-        public static AggregateBoundAttributeDescription GetAttributeDescriptionInfo(this CompletionItem completion)
-        {
-            if (completion is null)
-            {
-                throw new ArgumentNullException(nameof(completion));
-            }
-
-            if (completion.Data is JObject data && data.ContainsKey(AttributeCompletionDataKey))
-            {
-                var descriptionInfo = data[AttributeCompletionDataKey].ToObject<AggregateBoundAttributeDescription>();
-                return descriptionInfo;
-            }
-
-            return null;
         }
     }
 }