- We now only send completion item documentation when a completion item is resolved.
  - To accomplish this we now cache our Razor completion lists between responses and hold up to 3 in memory at a time. When a completion resolve request comes in we'll cross-check the completion item with our completion list cache and if we can find it we'll resolve/calculate the documentation information for that completion item.
- Expanded our pre-existing RazorCompletionEndpointTest's to encompass the new functionality
- Added separate CompletionListCache tests.

Baseline:
|                                              Method |      Mean |     Error |    StdDev |   Op/s |     Gen 0 |     Gen 1 |    Gen 2 | Allocated |
|---------------------------------------------------- |----------:|----------:|----------:|-------:|----------:|----------:|---------:|----------:|
| 'Component Completion List Roundtrip Serialization' | 29.436 ms | 0.5058 ms | 0.4731 ms |  33.97 | 2187.5000 | 1187.5000 | 250.0000 |   13.5 MB |
|           'Component Completion List Serialization' |  6.408 ms | 0.0853 ms | 0.0798 ms | 156.06 |  296.8750 |  242.1875 | 234.3750 |    2.3 MB |
|         'Component Completion List Deserialization' | 24.167 ms | 0.4655 ms | 0.6526 ms |  41.38 | 1968.7500 |  906.2500 | 125.0000 |  11.21 MB |

After:
|                                              Method |      Mean |     Error |    StdDev |   Op/s |     Gen 0 |    Gen 1 |   Gen 2 |  Allocated |
|---------------------------------------------------- |----------:|----------:|----------:|-------:|----------:|---------:|--------:|-----------:|
| 'Component Completion List Roundtrip Serialization' | 15.999 ms | 0.3149 ms | 0.4714 ms |  62.50 | 1031.2500 | 531.2500 | 31.2500 | 6514.11 KB |
|           'Component Completion List Serialization' |  2.301 ms | 0.0286 ms | 0.0268 ms | 434.66 |  128.9063 |  85.9375 | 62.5000 |  811.62 KB |
|         'Component Completion List Deserialization' | 11.451 ms | 0.0883 ms | 0.0690 ms |  87.33 |  921.8750 | 437.5000 |       - | 5702.63 KB |

*Roundtrip op/s improvement:* 1.84x faster
*Serialization op/s improvement:* 2.79x faster
*Deserialization op/s improvement:* 2.11x faster

Fixes dotnet/aspnetcore#29096